Updated pizzaSpawner settings properly. [NOTICE]
I have completely fixed the pizza prefab spawner. The only problem is, there needs to be an actual pizza object present in the scene for it to spawn. I need to figure out another way to have the spawner spawn the prefab without having one initialized in the map from the start.